+++ StarLevelSystem/modules/LootLevelsExpanded.cs
@@ -181,21 +181,29 @@ namespace StarLevelSystem.modules
             }
         }
 
+        internal static List<KeyValuePair<GameObject, int>> ModifyTreeDropsOrDefault(TreeBase treeInstance) {
+            int level = CompositeLazyCache.GetOrAddCachedTreeEntry(treeInstance.m_nview);
+            SelectObjectDistanceBonus(treeInstance.transform, out DistanceLootModifier distance_bonus);
+            List<KeyValuePair<GameObject, int>> drops = ModifyObjectDropsOrDefault(treeInstance.m_dropWhenDestroyed, Utils.GetPrefabName(treeInstance.gameObject), level, distance_bonus, DropType.Tree);
+            return drops;
+        }
+
         internal static List<KeyValuePair<GameObject, int>> ModifyTreeDropsOrDefault(TreeLog treeInstance) {
             int level = CompositeLazyCache.GetOrAddCachedTreeEntry(treeInstance.m_nview);
             SelectObjectDistanceBonus(treeInstance.transform, out DistanceLootModifier distance_bonus);
-            return ModifyObjectDropsOrDefault(treeInstance.m_dropWhenDestroyed, Utils.GetPrefabName(treeInstance.gameObject), level, distance_bonus, DropType.Tree);
+            List<KeyValuePair<GameObject, int>> drops = ModifyObjectDropsOrDefault(treeInstance.m_dropWhenDestroyed, Utils.GetPrefabName(treeInstance.gameObject), level, distance_bonus, DropType.Tree);
+            return drops;
         }
 
         internal static List<KeyValuePair<GameObject, int>> ModifyRockDropsOrDefault(Transform tform, DropTable droptable, string name, int level) {
             SelectObjectDistanceBonus(tform, out DistanceLootModifier distance_bonus);
-            return ModifyObjectDropsOrDefault(droptable, name, level, distance_bonus, DropType.Rock);
+            List<KeyValuePair<GameObject, int>> drops = ModifyObjectDropsOrDefault(droptable, name, level, distance_bonus, DropType.Rock);
+            return drops;
         }
 
 
         internal static List<KeyValuePair<GameObject, int>> ModifyObjectDropsOrDefault(DropTable defaultDrops, string lookupkey, int level, DistanceLootModifier distance_bonus, DropType type = DropType.Tree) {
             List<KeyValuePair<GameObject, int>> dropList = new List<KeyValuePair<GameObject, int>>();
-
             Logger.LogDebug($"Checking for custom drop configuration for {lookupkey}");
             if (LootSystemData.SLS_Drop_Settings != null && LootSystemData.SLS_Drop_Settings.nonCharacterSpecificLoot != null && LootSystemData.SLS_Drop_Settings.nonCharacterSpecificLoot.ContainsKey(lookupkey)) {
                 Logger.LogDebug($"Custom loot drops configured for:{lookupkey}");
@@ -261,6 +269,7 @@ namespace StarLevelSystem.modules
 
                     dropList.Add(new KeyValuePair<GameObject, int>(eod.DropGo, drop));
                 }
+                
             } else {
                 // generate the default loot list for this tree
                 if (defaultDrops == null) { return dropList; }
@@ -382,7 +391,7 @@ namespace StarLevelSystem.modules
                 // Modify Loot Drop for minerock5
                 List<KeyValuePair<GameObject, int>> optimizeDrops = ModifyRockDropsOrDefault(instance.transform, instance.m_dropItems, Utils.GetPrefabName(instance.gameObject), LevelSystem.DeterministicDetermineRockLevel(instance.gameObject.transform.position));
                 Vector3 position = hit.m_point - hit.m_dir * 0.2f + UnityEngine.Random.insideUnitSphere * 0.3f;
-                LootLevelsExpanded.DropItemsPreferAsync(position, optimizeDrops);
+                LootLevelsExpanded.DropItemsPreferAsync(position, optimizeDrops, dropThatNonCharacterDrop: true);
             }
         }
 
@@ -396,7 +405,7 @@ namespace StarLevelSystem.modules
             [HarmonyPatch(nameof(MineRock5.DamageArea))]
             static IEnumerable<CodeInstruction> Transpiler(IEnumerable<CodeInstruction> instructions, ILGenerator generator) {
                 var codeMatcher = new CodeMatcher(instructions, generator);
-                if (StarLevelSystem.IsDropThatEnabled) {
+                if (Compatibility.IsDropThatEnabled && Compatibility.DropThatMethodsAvailable == false) {
                     Logger.LogDebug("DropThat detected, using non-performance based Minerock5 patch for compat.");
                     codeMatcher
                         .MatchStartForward(
@@ -430,7 +439,7 @@ namespace StarLevelSystem.modules
             internal static void MineDrop(MineRock5 instance, Vector3 vector) {
                 int level = LevelSystem.DeterministicDetermineRockLevel(vector);
                 List<KeyValuePair<GameObject, int>> optimizeDrops = ModifyRockDropsOrDefault(instance.transform, instance.m_dropItems, Utils.GetPrefabName(instance.gameObject), level);
-                LootLevelsExpanded.DropItemsPreferAsync(instance.transform.position, optimizeDrops);
+                LootLevelsExpanded.DropItemsPreferAsync(instance.transform.position, optimizeDrops, dropThatNonCharacterDrop: true);
             }
             
             // This is specifically used for compatibility with DropThat, as removing the loop will break DropThats functionality for Minerock5
@@ -474,7 +483,7 @@ namespace StarLevelSystem.modules
                 int level = LevelSystem.DetermineisticDetermineObjectLevel(instance.transform.position);
                 SelectObjectDistanceBonus(instance.transform, out DistanceLootModifier distance_bonus);
                 List<KeyValuePair<GameObject, int>> optimizeDrops = ModifyObjectDropsOrDefault(instance.m_dropWhenDestroyed, Utils.GetPrefabName(instance.gameObject), level, distance_bonus, DropType.Destructible);
-                LootLevelsExpanded.DropItemsPreferAsync(instance.transform.position, optimizeDrops);
+                LootLevelsExpanded.DropItemsPreferAsync(instance.transform.position, optimizeDrops, dropThatNonCharacterDrop: true);
             }
         }
 
@@ -487,29 +496,54 @@ namespace StarLevelSystem.modules
             [HarmonyPatch(nameof(TreeBase.RPC_Damage))]
             static IEnumerable<CodeInstruction> Transpiler(IEnumerable<CodeInstruction> instructions /*, ILGenerator generator*/) {
                 var codeMatcher = new CodeMatcher(instructions);
-                codeMatcher.MatchStartForward(
-                    new CodeMatch(OpCodes.Ldarg_0),
-                    new CodeMatch(OpCodes.Ldfld),
-                    new CodeMatch(OpCodes.Callvirt, AccessTools.Method(typeof(DropTable), nameof(DropTable.GetDropList)))
-                    ).Advance(1)
-                    .InsertAndAdvance(
-                        Transpilers.EmitDelegate(TreebaseDropDestroyedItems)
-                    )
-                    //.MatchStartForward(
-                    //new CodeMatch(OpCodes.)
-                    .RemoveInstructions(54)
-                    .Insert(
-                        new CodeInstruction(OpCodes.Ldarg_0)
-                    )
-                    .ThrowIfNotMatch("Unable to patch Treebase to handle large drops.");
+
+                if (Compatibility.IsDropThatEnabled && Compatibility.DropThatMethodsAvailable == false) {
+                    Logger.LogDebug("DropThat detected, using non-performance based TreeBase patch for compat.");
+                    codeMatcher
+                        .MatchStartForward(
+                            new CodeMatch(OpCodes.Ldarg_0),
+                            new CodeMatch(OpCodes.Ldfld),
+                            new CodeMatch(OpCodes.Callvirt, AccessTools.Method(typeof(DropTable), nameof(DropTable.GetDropList)))
+                            )
+                        .Advance(1) // includes the Ldarg_0
+                        .RemoveInstructions(2)
+                        .InsertAndAdvance(
+                            Transpilers.EmitDelegate(NonPerformanceBasedTreeBaseDrop)
+                        ).ThrowIfNotMatch("Unable to patch TreeBase to provide loot modifications.");
+                } else {
+                    codeMatcher.MatchStartForward(
+                        new CodeMatch(OpCodes.Ldarg_0),
+                        new CodeMatch(OpCodes.Ldfld),
+                        new CodeMatch(OpCodes.Callvirt, AccessTools.Method(typeof(DropTable), nameof(DropTable.GetDropList)))
+                        ).Advance(1)
+                        .InsertAndAdvance(
+                            Transpilers.EmitDelegate(TreebaseDropDestroyedItems)
+                        )
+                        //.MatchStartForward(
+                        //new CodeMatch(OpCodes.)
+                        .RemoveInstructions(54)
+                        .Insert(
+                            new CodeInstruction(OpCodes.Ldarg_0)
+                        )
+                        .ThrowIfNotMatch("Unable to patch Treebase to handle large drops.");
+                }
+
                 return codeMatcher.Instructions();
             }
 
             private static void TreebaseDropDestroyedItems(TreeBase instance) {
-                int level = LevelSystem.DetermineisticDetermineObjectLevel(instance.transform.position);
-                SelectObjectDistanceBonus(instance.transform, out DistanceLootModifier distance_bonus);
-                List<KeyValuePair<GameObject, int>> optimizeDrops = ModifyObjectDropsOrDefault(instance.m_dropWhenDestroyed, Utils.GetPrefabName(instance.gameObject), level, distance_bonus, DropType.Destructible);
-                LootLevelsExpanded.DropItemsPreferAsync(instance.transform.position, optimizeDrops);
+                List<KeyValuePair<GameObject, int>> optimizeDrops = ModifyTreeDropsOrDefault(instance);
+                LootLevelsExpanded.DropItemsPreferAsync(instance.transform.position, optimizeDrops, dropThatNonCharacterDrop: true);
+            }
+
+            internal static List<GameObject> NonPerformanceBasedTreeBaseDrop(TreeBase instance) {
+                // Modify Loot Drop for TreeBase
+                List<GameObject> drops = new List<GameObject>();
+                List<KeyValuePair<GameObject, int>> optimizeDrops = ModifyTreeDropsOrDefault(instance);
+                foreach (KeyValuePair<GameObject, int> drop in optimizeDrops) {
+                    drop.Value.Times(() => drops.Add(drop.Key));
+                }
+                return drops;
             }
         }
 
@@ -518,12 +552,7 @@ namespace StarLevelSystem.modules
             // effectively replace the drop items function since we need to drop things in a way that is not insane for large amounts of loot
             [HarmonyPatch(nameof(CharacterDrop.DropItems))]
             public static bool Prefix(CharacterDrop __instance, List<KeyValuePair<GameObject, int>> drops, Vector3 centerPos, float dropArea) {
-                if (Player.m_localPlayer != null) {
-                    Player.m_localPlayer.StartCoroutine(DropItemsAsync(drops, centerPos, dropArea));
-                } else {
-                    DropItemsImmediate(drops, centerPos, dropArea);
-                }
-
+                DropItemsPreferAsync(centerPos, drops, dropThatCharacterDrop: true);
                 return false;
             }
         }
@@ -553,20 +582,16 @@ namespace StarLevelSystem.modules
             return optimizeDrops;
         }
 
-        internal static void DropItems(Vector3 position, List<GameObject> drops, bool immediate = false, float lootmult = 1) {
-            List<KeyValuePair<GameObject, int>> optimizeDrops = OptimizeToDropStacks(drops, lootmult);
-            DropItemsPreferAsync(position, optimizeDrops, immediate);
-        }
-
-        public static void DropItemsPreferAsync(Vector3 position, List<KeyValuePair<GameObject, int>> optimizeDrops, bool immediate = false) {
+        public static void DropItemsPreferAsync(Vector3 position, List<KeyValuePair<GameObject, int>> optimizeDrops, bool immediate = false, bool dropThatCharacterDrop = false, bool dropThatNonCharacterDrop = false) {
             if (Player.m_localPlayer != null && immediate == false) {
-                Player.m_localPlayer.StartCoroutine(DropItemsAsync(optimizeDrops, position, 0.5f));
+                Player.m_localPlayer.StartCoroutine(DropItemsAsync(optimizeDrops, position, 0.5f, dropThatCharacterDrop));
             } else {
-                DropItemsImmediate(optimizeDrops, position, 0.5f);
+                DropItemsImmediate(optimizeDrops, position, 0.5f, dropThatCharacterDrop, dropThatNonCharacterDrop);
             }
         }
 
-        internal static void DropItemsImmediate(List<KeyValuePair<GameObject, int>> drops, Vector3 centerPos, float dropArea) {
+        internal static void DropItemsImmediate(List<KeyValuePair<GameObject, int>> drops, Vector3 centerPos, float dropArea, bool dropThatCharacterDrop = false, bool dropThatNonCharacterDrop = false) {
+            int dropindex = 0;
             foreach (var drop in drops) {
                 bool set_stack_size = false;
                 int max_stack_size = 0;
@@ -607,6 +632,15 @@ namespace StarLevelSystem.modules
                         }
                     }
 
+                    // Compat for DropThat
+                    // Send the dropped items to drop that to allow modifications
+                    if (dropThatCharacterDrop && Compatibility.IsDropThatEnabled && Compatibility.DropThatMethodsAvailable) {
+                        Compatibility.DropThat_ModifyDrop(droppedItem, drops, dropindex);
+                    }
+                    if (dropThatNonCharacterDrop && Compatibility.IsDropThatEnabled && Compatibility.DropThatMethodsAvailable) {
+                        Compatibility.DropThat_ModifyInstantiatedObjectDrop(droppedItem, dropindex);
+                    }
+
                     Rigidbody component2 = droppedItem.GetComponent<Rigidbody>();
                     if ((bool)component2) {
                         Vector3 insideUnitSphere = UnityEngine.Random.insideUnitSphere * dropArea;
@@ -617,14 +651,15 @@ namespace StarLevelSystem.modules
                     }
                     i++;
                 }
+                dropindex++;
             }
         }
 
         
 
-        public static IEnumerator DropItemsAsync(List<KeyValuePair<GameObject, int>> drops, Vector3 centerPos, float dropArea) {
+        public static IEnumerator DropItemsAsync(List<KeyValuePair<GameObject, int>> drops, Vector3 centerPos, float dropArea, bool dropThatCharacterDrop = false, bool dropThatNonCharacterDrop = false) {
             int obj_spawns = 0;
-
+            int dropindex = 0;
             foreach (var drop in drops)
             {
                 bool set_stack_size = false;
@@ -679,7 +714,17 @@ namespace StarLevelSystem.modules
                             ModificationExtensionSystem.CreatureSetup(chara, delay: 0.5f);
                         }
                     }
-                        Rigidbody component2 = droppedItem.GetComponent<Rigidbody>();
+
+                    // Compat for DropThat
+                    // Send the dropped items to drop that to allow modifications
+                    if (dropThatCharacterDrop && Compatibility.IsDropThatEnabled && Compatibility.DropThatMethodsAvailable) {
+                        Compatibility.DropThat_ModifyDrop(droppedItem, drops, dropindex);
+                    }
+                    if (dropThatNonCharacterDrop && Compatibility.IsDropThatEnabled && Compatibility.DropThatMethodsAvailable) {
+                        Compatibility.DropThat_ModifyInstantiatedObjectDrop(droppedItem, dropindex);
+                    }
+
+                    Rigidbody component2 = droppedItem.GetComponent<Rigidbody>();
                     if ((bool)component2) {
                         Vector3 insideUnitSphere = UnityEngine.Random.insideUnitSphere * dropArea;
                         if (insideUnitSphere.y < 0f) {
@@ -689,6 +734,7 @@ namespace StarLevelSystem.modules
                     }
                     i++;
                 }
+                dropindex++;
             }
 
             yield break;