+++ StarLevelSystem/Package/CHANGELOG.md
@@ -1,3 +1,21 @@
+**0.17.3**
+ ---
+ ```
+ - Improves performance when used with DropThat
+ ```
+
+**0.17.2**
+ ---
+ ```
+ - Health coalesence multiplication fix
+ ```
+
+**0.17.1**
+ ---
+ ```
+ - Improves compatibility with Drop That
+ ```
+
 **0.17.0**
  ---
  ```